# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0",
      "title": "swagger",
      "contact": {}
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "securityDefinitions": {},
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/user/{userId}": {
        "put": {
          "summary": "update user",
          "tags": [
            "user"
          ],
          "operationId": "updateuser",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Body",
              "in": "body",
              "required": true,
              "description": "",
              "schema": {
                "$ref": "#/definitions/updateuserrequest"
              }
            },
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        }
      },
      "/products": {
        "get": {
          "summary": "Get one product",
          "tags": [
            "products"
          ],
          "operationId": "Getoneproduct",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        },
        "post": {
          "summary": "create product",
          "tags": [
            "products"
          ],
          "operationId": "createproduct",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Body",
              "in": "body",
              "required": true,
              "description": "",
              "schema": {
                "$ref": "#/definitions/createproductrequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        },
        "put": {
          "summary": "update Products",
          "tags": [
            "products"
          ],
          "operationId": "updateProducts",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        },
        "delete": {
          "summary": "delete Products",
          "tags": [
            "products"
          ],
          "operationId": "deleteProducts",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        }
      }
    },
    "definitions": {
...
      "updateuserrequest": {
        "title": "updateuserrequest",
        "example": {
          "userName": "shubham savaliya",
          "email": "shubham@gmail.com",
          "password": "viral123",
          "age": 21,
          "address": "surat"
        },
        "type": "object",
        "properties": {
          "userName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "type": "string"
          }
        },
        "required": [
          "userName",
          "email",
          "password",
          "age",
          "address"
        ]
      },
      "createproductrequest": {
        "title": "createproductrequest",
        "example": {
          "productName": "t.v",
          "price": "30,000"
        },
        "type": "object",
        "properties": {
          "productName": {
            "type": "string"
          },
          "price": {
            "type": "string"
          }
        },
        "required": [
          "productName",
          "price"
        ]
      }
    },
...
  }